Make typing more straightforward
